# Backend/src/main/java/com/hogaria/expense/dto/ExpenseDto.java
package com.hogaria.expense.dto;

import java.math.BigDecimal;
import java.time.LocalDate;

public record ExpenseDto(
        Long id,
        Long familyId,
        Long userId,
        BigDecimal amount,
        String currency,
        Long categoryId,
        Long subcategoryId,
        String descripcion,
        LocalDate date
) {}