# Backend/src/main/java/com/hogaria/expense/dto/RecurringExpenseDto.java
package com.hogaria.expense.dto;


import java.time.LocalDate;

public record RecurringExpenseDto(
        Long id,
        Long expenseTemplateId,
        String frequency,
        LocalDate nextDueDate
) {
}