# Backend/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Gestion Economica e Inventario del Hogar API
  version: 1.0.0
  description: API RESTful para gestionar usuarios, casas, familias, inventario, gastos y notificaciones.
servers:
  - url: https://api.tuhogar.app/v1
    description: Servidor de produccion
  - url: http://localhost:8080/v1
    description: Servidor de desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        expires_in:
          type: integer
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
    HouseDto:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        direccion:
          type: string
    FamilyDto:
      type: object
      properties:
        id:
          type: integer
        house_id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
...
    ExpenseDto:
      type: object
      properties:
        id:
          type: integer
        family_id:
          type: integer
        user_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
        category_id:
          type: integer
        subcategory_id:
          type: integer
        descripcion:
          type: string
        date:
          type: string
          format: date
    RecurringExpenseDto:
      type: object
      properties:
        id:
          type: integer
        expense_template_id:
          type: integer
        frequency:
          type: string
          enum:
            - DAILY
            - WEEKLY
            - BIWEEKLY
            - MONTHLY
            - QUARTERLY
            - SEMIANNUAL
            - ANNUAL
        next_due_date:
          type: string
          format: date
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Registro de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          description: Bad Request
        '409':
          description: Conflict - usuario existente
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '200':
          description: Autenticacion exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
...
  /houses:
    get:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Listar casas
      responses:
        '200':
          description: Lista de casas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseDto'
    post:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Crear casa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                direccion:
                  type: string
      responses:
        '201':
          description: Casa creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseDto'
  /houses/{house_id}/families:
    get:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Listar familias en casa
      parameters:
        - in: path
          name: house_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Lista de familias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyDto'
        '404':
          description: Casa no encontrada
    post:
      tags:
        - House
      security:
        - bearerAuth: [ ]
      summary: Crear familia
      parameters:
        - in: path
          name: house_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Familia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyDto'
...
  /expenses:
    get:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Listar gastos
      parameters:
        - in: query
          name: family_id
          schema:
            type: integer
          required: true
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: category_id
          schema:
            type: integer
        - in: query
          name: subcategory_id
          schema:
            type: integer
      responses:
        '200':
          description: Lista de gastos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseDto'
    post:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Crear gasto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseDto'
      responses:
        '201':
          description: Gasto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseDto'
  /expenses/{id}:
    put:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Actualizar gasto
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseDto'
      responses:
        '200':
          description: Gasto actualizado
  /recurring-expenses:
    get:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Listar gastos recurrentes
      parameters:
        - in: query
          name: family_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Lista de gastos recurrentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringExpenseDto'
    post:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Crear gasto recurrente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseDto'
      responses:
        '201':
          description: Gasto recurrente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpenseDto'
  /recurring-expenses/{id}:
    put:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Actualizar gasto recurrente
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                next_due_date:
                  type: string
                  format: date
      responses:
        '200':
          description: Gasto recurrente actualizado
    delete:
      tags:
        - Expense
      security:
        - bearerAuth: [ ]
      summary: Eliminar gasto recurrente
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: Gasto recurrente eliminado
...